+++ packages/builder-tools/connect-studio/vite-plugins/external-packages.ts
@@ -1,7 +1,7 @@
 import { getConfig } from "@powerhousedao/config/powerhouse";
 import fs from "node:fs";
-import { dirname, join } from "node:path";
-import { normalizePath, PluginOption } from "vite";
+import { join } from "node:path";
+import { PluginOption } from "vite";
 import { viteIgnoreStaticImport, viteReplaceImports } from "./base.js";
 
 // TODO use config path?
@@ -9,17 +9,14 @@ const __dirname = join(process.cwd(), ".ph/");
 // import.meta.dirname || dirname(fileURLToPath(import.meta.url));
 
 export const EXTERNAL_PACKAGES_IMPORT = "PH:EXTERNAL_PACKAGES";
-export const IMPORT_SCRIPT_FILE = normalizePath(
-  join(__dirname, "external-packages.js"),
-);
+export const IMPORT_SCRIPT_FILE = "external-packages.js";
 
-export function generateImportScript(packages: string[]) {
-  const targetDir = dirname(IMPORT_SCRIPT_FILE);
+function generateImportScript(packages: string[], targetDir: string) {
   // create file if it doesn't exist, also create path if it doesn't exist (recursive)
   if (!fs.existsSync(targetDir)) {
     fs.mkdirSync(targetDir, { recursive: true });
   }
-
+  const scriptPath = join(targetDir, IMPORT_SCRIPT_FILE);
   const imports: string[] = [];
   const moduleNames: string[] = [];
   let counter = 0;
@@ -43,23 +40,25 @@ export function generateImportScript(packages: string[]) {
     ];`;
 
   const fileContent = `${imports.join("\n")}\n\n${exportStatement}`;
-  fs.writeFileSync(IMPORT_SCRIPT_FILE, fileContent);
+  fs.writeFileSync(scriptPath, fileContent);
 
-  return IMPORT_SCRIPT_FILE;
+  return scriptPath;
 }
 
 export const viteLoadExternalPackages = (
   packages: string[] | undefined,
+  targetDir: string,
   hmr = false,
 ): PluginOption => {
   if (!packages?.length && !hmr) {
     return viteIgnoreStaticImport([EXTERNAL_PACKAGES_IMPORT]);
   }
 
-  generateImportScript(packages ?? []);
+  const importScriptPath = generateImportScript(packages ?? [], targetDir);
+
   process.env.LOAD_EXTERNAL_PACKAGES = "true";
   return [
-    viteReplaceImports({ [EXTERNAL_PACKAGES_IMPORT]: IMPORT_SCRIPT_FILE }),
+    viteReplaceImports({ [EXTERNAL_PACKAGES_IMPORT]: importScriptPath }),
     hmr && {
       name: "vite-plugin-studio-external-packages",
       handleHotUpdate({ file, server, modules }) {
@@ -67,6 +66,7 @@ export const viteLoadExternalPackages = (
           const config = getConfig(file);
           generateImportScript(
             config.packages?.map((p) => p.packageName) ?? [],
+            targetDir,
           );
 
           config.packages?.forEach((pkg) =>