+++ packages/codegen/src/codegen/.hygen/templates/powerhouse/generate-document-model-module/creators.esm.t
@@ -2,7 +2,7 @@
 to: "<%= rootDir %>/<%= h.changeCase.param(documentType) %>/gen/<%= module %>/creators.ts"
 force: true
 ---
-import { utils<% if (actions.find(a => a.hasAttachment)) {%>, AttachmentInput<%}%> } from 'document-model/document';
+import { createAction<% if (actions.find(a => a.hasAttachment)) {%>, AttachmentInput<%}%> } from 'document-model';
 import { z,
 <% actions.filter(a => a.hasInput).forEach(action => { _%>
     <%= h.changeCase.pascal(action.name) %>Input,
@@ -13,8 +13,6 @@ import {
     <%= h.changeCase.pascal(action.name) %>Action,
 <% }); _%>
 } from './actions';
-
-const { createAction } = utils;
 
 <% actions.filter(a => a.hasInput).forEach(action => { _%>
 export const <%= h.changeCase.camel(action.name) %> = (input: <%= h.changeCase.pascal(action.name) %>Input<%if(action.hasAttachment){ %>, attachments: AttachmentInput[] <% } %>) =>