+++ packages/document-drive/test/server/processOperations.test.ts
@@ -1,7 +1,9 @@
 import {
   Action,
   DocumentModelDocument,
-  DocumentModelLib,
+  documentModelDocumentModelModule,
+  DocumentModelModule,
+  documentModelReducer,
   garbageCollect,
   Operation,
   setModelExtension,
@@ -9,19 +11,17 @@ import {
   setModelName,
 } from "document-model";
 import { beforeEach, describe, expect, it } from "vitest";
-import { IOperationResult, ReactorBuilder } from "../../src";
-import { OperationError } from "../../src/server/error";
-import { garbageCollect } from "../../src/utils/document-helpers";
-import { BasicClient, buildOperation, buildOperations } from "../utils";
 
 import { undo } from "../../../document-model/src/document/actions/creators.js";
 import { DocumentDriveAction } from "../../src/drive-document-model/gen/actions.js";
-import { reducer } from "../../src/drive-document-model/gen/reducer.js";
+import { reducer as documentDriveReducer } from "../../src/drive-document-model/gen/reducer.js";
+import { driveDocumentModelModule } from "../../src/drive-document-model/module.js";
 import { generateAddNodeAction } from "../../src/drive-document-model/src/utils.js";
-import { DocumentDriveServer } from "../../src/server/base.js";
+import { ReactorBuilder } from "../../src/server/builder.js";
 import { OperationError } from "../../src/server/error.js";
 import { IOperationResult } from "../../src/server/types.js";
 import { BasicClient, buildOperation, buildOperations } from "../utils.js";
+
 const mapExpectedOperations = (operations: Operation[]) =>
   operations.map((op) => {
     const { timestamp, ...operation } = op;
@@ -30,9 +30,9 @@ const mapExpectedOperations = (operations: Operation[]) =>
 
 describe("processOperations", () => {
   const documentModels = [
-    DocumentModelLib,
-    ...Object.values(DocumentModelsLibs),
-  ] as DocumentModelLib<any, any>[];
+    documentModelDocumentModelModule,
+    driveDocumentModelModule,
+  ] as DocumentModelModule[];
 
   let server = new ReactorBuilder(documentModels).build();
   beforeEach(async () => {
@@ -57,7 +57,7 @@ describe("processOperations", () => {
     await server.addDriveOperation(
       driveId,
       buildOperation(
-        reducer,
+        documentDriveReducer,
         drive,
         generateAddNodeAction(
           drive.state.global,
@@ -80,7 +80,7 @@ describe("processOperations", () => {
       await server.addOperations(
         driveId,
         documentId,
-        buildOperations(reducer, document, initialOperations),
+        buildOperations(documentModelReducer, document, initialOperations),
       );
 
       document = (await server.getDocument(
@@ -110,7 +110,7 @@ describe("processOperations", () => {
   it("should apply a single new operation", async () => {
     const document = await buildFile();
 
-    const operations = buildOperations(reducer, document, [
+    const operations = buildOperations(documentModelReducer, document, [
       setModelName({ name: "test" }),
     ]);
 
@@ -135,7 +135,7 @@ describe("processOperations", () => {
       setModelId({ id: "test" }),
     ]);
 
-    const operations = buildOperations(reducer, document, [
+    const operations = buildOperations(documentModelReducer, document, [
       setModelName({ name: "test2" }),
       setModelId({ id: "test2" }),
       setModelExtension({
@@ -168,7 +168,9 @@ describe("processOperations", () => {
       setModelId({ id: "test" }),
     ]);
 
-    const operations = buildOperations(reducer, document, [undo()]);
+    const operations = buildOperations(documentModelReducer, document, [
+      undo(),
+    ]);
 
     const result = await server._processOperations(
       driveId,
@@ -206,7 +208,9 @@ describe("processOperations", () => {
       undo(),
     ]);
 
-    const operations = buildOperations(reducer, document, [undo()]);
+    const operations = buildOperations(documentModelReducer, document, [
+      undo(),
+    ]);
 
     const result = await server._processOperations(
       driveId,
@@ -240,7 +244,12 @@ describe("processOperations", () => {
     ]);
 
     const operations = [
-      buildOperation(reducer, document, setModelName({ name: "test2" }), 4),
+      buildOperation(
+        documentModelReducer,
+        document,
+        setModelName({ name: "test2" }),
+        4,
+      ),
     ];
 
     const result = await server._processOperations(
@@ -271,10 +280,30 @@ describe("processOperations", () => {
     ]);
 
     const operations = [
-      buildOperation(reducer, document, setModelName({ name: "test3" }), 3),
-      buildOperation(reducer, document, setModelName({ name: "test4" }), 4),
-      buildOperation(reducer, document, setModelName({ name: "test6" }), 6),
-      buildOperation(reducer, document, setModelName({ name: "test7" }), 7),
+      buildOperation(
+        documentModelReducer,
+        document,
+        setModelName({ name: "test3" }),
+        3,
+      ),
+      buildOperation(
+        documentModelReducer,
+        document,
+        setModelName({ name: "test4" }),
+        4,
+      ),
+      buildOperation(
+        documentModelReducer,
+        document,
+        setModelName({ name: "test6" }),
+        6,
+      ),
+      buildOperation(
+        documentModelReducer,
+        document,
+        setModelName({ name: "test7" }),
+        7,
+      ),
     ];
 
     const result = await server._processOperations(
@@ -305,7 +334,12 @@ describe("processOperations", () => {
     ]);
 
     const operations = [
-      buildOperation(reducer, document, setModelName({ name: "test2" }), 2),
+      buildOperation(
+        documentModelReducer,
+        document,
+        setModelName({ name: "test2" }),
+        2,
+      ),
     ];
 
     const result = await server._processOperations(
@@ -333,9 +367,24 @@ describe("processOperations", () => {
     ]);
 
     const operations = [
-      buildOperation(reducer, document, setModelName({ name: "test3" }), 3),
-      buildOperation(reducer, document, setModelName({ name: "test4" }), 3),
-      buildOperation(reducer, document, setModelName({ name: "test5" }), 4),
+      buildOperation(
+        documentModelReducer,
+        document,
+        setModelName({ name: "test3" }),
+        3,
+      ),
+      buildOperation(
+        documentModelReducer,
+        document,
+        setModelName({ name: "test4" }),
+        3,
+      ),
+      buildOperation(
+        documentModelReducer,
+        document,
+        setModelName({ name: "test5" }),
+        4,
+      ),
     ];
 
     const result = await server._processOperations(
@@ -359,7 +408,7 @@ describe("processOperations", () => {
     let document = await buildFile();
 
     const operation = buildOperation(
-      reducer,
+      documentModelReducer,
       document,
       setModelName({ name: "test" }),
     );
@@ -411,7 +460,7 @@ describe("processOperations", () => {
     let document = await buildFile();
 
     const operation0 = buildOperation(
-      reducer,
+      documentModelReducer,
       document,
       setModelName({ name: "1" }),
       0,
@@ -431,7 +480,7 @@ describe("processOperations", () => {
     )) as DocumentModelDocument;
 
     const operation1 = buildOperation(
-      reducer,
+      documentModelReducer,
       document,
       setModelName({ name: "2" }),
       0,
@@ -451,7 +500,7 @@ describe("processOperations", () => {
     )) as DocumentModelDocument;
 
     const operation2 = buildOperation(
-      reducer,
+      documentModelReducer,
       document,
       setModelId({ id: "3" }),
       0,
@@ -471,7 +520,7 @@ describe("processOperations", () => {
     )) as DocumentModelDocument;
 
     const operation3 = buildOperation(
-      reducer,
+      documentModelReducer,
       document,
       setModelId({ id: "4" }),
       0,
@@ -533,7 +582,7 @@ describe("processOperations", () => {
       driveId,
       documentId,
       initialDocument,
-      reducer,
+      documentModelReducer,
     );
 
     const client2 = new BasicClient(
@@ -541,7 +590,7 @@ describe("processOperations", () => {
       driveId,
       documentId,
       initialDocument,
-      reducer,
+      documentModelReducer,
     );
 
     client1.dispatchDocumentAction(setModelName({ name: "1" }));
@@ -609,7 +658,7 @@ describe("processOperations", () => {
       driveId,
       documentId,
       initialDocument,
-      reducer,
+      documentModelReducer,
     );
 
     const client2 = new BasicClient(
@@ -617,7 +666,7 @@ describe("processOperations", () => {
       driveId,
       documentId,
       initialDocument,
-      reducer,
+      documentModelReducer,
     );
 
     client1.dispatchDocumentAction(setModelName({ name: "1" }));